requirements_gui.txt
# --- Core app ---
streamlit>=1.33,<2.0
pandas>=2.1,<3.0
numpy>=1.26,<3.0

# Walidacja i daty
pydantic>=2.6,<3.0
python-dateutil>=2.8.2

# Przydatne w dev/Windows (hot-reload)
watchdog>=4.0.0

# (opcjonalne, ale przydatne pod wykresy)
plotly>=5.20.0